scatter: fix comments
