+++ tests/ttnn/nightly/unit_tests/operations/data_movement/test_scatter.py
