+++ ttnn/cpp/ttnn/operations/data_movement/scatter/device/kernels/dataflow/reader_bf16_reduction_scatter.cpp
@@ -10,88 +10,11 @@
 
 namespace {
 
-template <int32_t N>
-FORCE_INLINE std::array<uint32_t, N> make_strides(const std::array<uint32_t, N>& dims) {
-    std::array<uint32_t, N> s{};
-    uint32_t acc = 1;
-    for (int32_t i = N - 1; i >= 0; --i) {
-        s[i] = acc;
-        acc *= dims[i];
-    }
-    return s;
-}
-
-template <int32_t N>
-FORCE_INLINE bool in_bounds(const std::array<uint32_t, N>& idx, const std::array<uint32_t, N>& dims) {
-    for (int32_t i = 0; i < N; ++i) {
-        if (idx[i] >= dims[i]) {
-            return false;
-        }
-    }
-    return true;
-}
-
-template <int32_t N>
-FORCE_INLINE bool next_inplace(std::array<uint32_t, N>& idx, const std::array<uint32_t, N>& dims) {
-    // last axis fastest
-    for (int32_t i = N - 1; i >= 0; --i) {
-        if (++idx[i] < dims[i]) {
-            return true;  // normal increment without carry
-        }
-        idx[i] = 0;  // carry and continue
-    }
-    return false;  // overflow past most significant digit
-}
-
-template <int32_t N>
-FORCE_INLINE uint32_t to_id(const std::array<uint32_t, N>& idx, const std::array<uint32_t, N>& strides) {
-    uint32_t id = 0;
-    for (int32_t i = 0; i < static_cast<int32_t>(N); ++i) {
-        id += idx[i] * strides[i];
-    }
-    return id;
-}
-
-// Convert linear id -> coordinates (row-major, last axis fastest).
-template <int32_t N>
-std::array<uint32_t, N> from_id(int32_t id, const std::array<uint32_t, N>& dims) {
-    std::array<uint32_t, N> coord{};
-    // Go left to right: for [d0, d1, ..., dN-1], last axis fastest
-    for (int32_t i = N - 1; i >= 0; --i) {
-        coord[i] = id % dims[i];
-        id /= dims[i];
-    }
-    return coord;
-}
-
-// this function is supposed to load either a whole stick or part of it
-template <typename AddrGen>
-FORCE_INLINE void load_to_cb(
-    uint32_t cb, const AddrGen& addr_gtor, uint32_t offset_bytes, uint32_t chunk_size_bytes, uint32_t stick_id) {
-    cb_reserve_back(cb, ONE_PAGE);
-    const uint64_t source_noc_address = get_noc_addr(stick_id, addr_gtor);
-    const uint32_t l1_write_address = get_write_ptr(cb);
-
-    noc_async_read(source_noc_address + offset_bytes, l1_write_address, chunk_size_bytes);
-    noc_async_read_barrier();
-
-    cb_push_back(cb, ONE_PAGE);
-}
-
 FORCE_INLINE static float bfloat16_to_float(uint16_t bfloat_val) {
     uint32_t uint32_data = ((uint32_t)bfloat_val) << 16;
     float f;
     std::memcpy(&f, &uint32_data, sizeof(f));
     return f;
-}
-
-FORCE_INLINE static uint16_t float_to_bfloat16(float val) {
-    union {
-        float f;
-        uint32_t u;
-    } ret;
-    ret.f = val;
-    return uint16_t(ret.u >> 16);
 }
 
 FORCE_INLINE std::uint16_t fp32_to_bf16(float x) {
