+++ ttnn/cpp/ttnn/operations/data_movement/scatter/device/kernels/dataflow/reader_scatter.cpp
@@ -10,78 +10,6 @@
 
 namespace {
 
-template <int32_t N>
-FORCE_INLINE std::array<uint32_t, N> make_strides(const std::array<uint32_t, N>& dims) {
-    std::array<uint32_t, N> s{};
-    uint32_t acc = 1;
-    for (int32_t i = N - 1; i >= 0; --i) {
-        s[i] = acc;
-        acc *= dims[i];
-    }
-    return s;
-}
-
-template <int32_t N>
-FORCE_INLINE bool in_bounds(const std::array<uint32_t, N>& idx, const std::array<uint32_t, N>& dims) {
-    for (int32_t i = 0; i < N; ++i) {
-        if (idx[i] >= dims[i]) {
-            return false;
-        }
-    }
-    return true;
-}
-
-template <int32_t N>
-FORCE_INLINE bool next_inplace(std::array<uint32_t, N>& idx, const std::array<uint32_t, N>& dims) {
-    // last axis fastest
-    for (int32_t i = N - 1; i >= 0; --i) {
-        if (++idx[i] < dims[i]) {
-            return true;  // normal increment without carry
-        }
-        idx[i] = 0;  // carry and continue
-    }
-    return false;  // overflow past most significant digit
-}
-
-template <int32_t N>
-FORCE_INLINE uint32_t to_id(const std::array<uint32_t, N>& idx, const std::array<uint32_t, N>& strides) {
-    uint32_t id = 0;
-    for (int32_t i = 0; i < static_cast<int32_t>(N); ++i) {
-        id += idx[i] * strides[i];
-    }
-    return id;
-}
-
-// Convert linear id -> coordinates (row-major, last axis fastest).
-template <int32_t N>
-std::array<uint32_t, N> from_id(int32_t id, const std::array<uint32_t, N>& dims) {
-    std::array<uint32_t, N> coord{};
-    // Go left to right: for [d0, d1, ..., dN-1], last axis fastest
-    for (int32_t i = N - 1; i >= 0; --i) {
-        coord[i] = id % dims[i];
-        id /= dims[i];
-    }
-    return coord;
-}
-
-// this function is supposed to load either a whole stick or part of it
-template <typename AddrGen>
-FORCE_INLINE void load_to_cb(
-    const uint32_t& cb,
-    const AddrGen& addr_gtor,
-    const uint32_t& offset_bytes,
-    const uint32_t& chunk_size_bytes,
-    const uint32_t& stick_id) {
-    cb_reserve_back(cb, ONE_PAGE);
-    const uint64_t source_noc_address = get_noc_addr(stick_id, addr_gtor);
-    const uint32_t l1_write_address = get_write_ptr(cb);
-
-    noc_async_read(source_noc_address + offset_bytes, l1_write_address, chunk_size_bytes);
-    noc_async_read_barrier();
-
-    cb_push_back(cb, ONE_PAGE);
-}
-
 // copies source stick to destination stick (first phase of scatter)
 template <typename number_type>
 FORCE_INLINE void copy_input_to_output(
