+++ ttnn/cpp/ttnn/operations/data_movement/scatter/device/scatter_program_factory.cpp
@@ -4,6 +4,8 @@
 
 #include "scatter_program_factory.hpp"
 
+#include "scatter_common.hpp"
+
 #include "scatter_device_operation_types.hpp"
 #include "tt-metalium/allocator.hpp"
 #include "tt-metalium/device.hpp"
@@ -13,61 +15,8 @@
 
 namespace ttnn::operations::data_movement::scatter {
 
-namespace {
 using namespace tt;
 using namespace tt::tt_metal;
-constexpr uint32_t BIT_MASK_32 = 32 - 1;
-
-uint64_t ceil32(const uint64_t& number) {
-    return ((number & BIT_MASK_32) == 0) ? number : ((number | BIT_MASK_32) + 1);
-}
-
-}  // namespace
-
-inline uint32_t get_max_l1_space(IDevice* device) {
-    auto lowest_address = device->lowest_occupied_compute_l1_address();
-    uint32_t max_l1_space = lowest_address.has_value() ? lowest_address.value() : device->l1_size_per_core();
-    max_l1_space = max_l1_space - device->allocator()->get_base_allocator_addr(HalMemType::L1);
-    return max_l1_space;
-}
-
-// maximal input/index/source/output chunk size, divisible by 32, calculated as follows:
-// BH available L1 mem size of nearly 1.5 MB...
-// ... divided by 4 to be able to allocate four equally long row chunks (coming from input/index/source/output
-// tensors)
-// ... divided by 4 to account for 4-byte datum sizes of each tensor (fp32, int32)
-// ... minimized by ~10% to account for reserved memory
-inline uint32_t calculate_optimal_chunk_size(IDevice* device) {
-    return ceil32(((((get_max_l1_space(device)) / 4) / 4) * 0.9) - 32);
-}
-
-CBHandle create_cb(
-    Program& program,
-    const DataType& dtype,
-    const ScatterCB& scatter_cb,
-    const CoreRangeSet& core_range_set,
-    const uint32_t& page_size_bytes) {
-    const uint32_t cb_id{static_cast<uint32_t>(scatter_cb)};
-    const auto cb_data_format{datatype_to_dataformat_converter(dtype)};
-    const auto cb_config{
-        CircularBufferConfig{page_size_bytes, {{cb_id, cb_data_format}}}.set_page_size(cb_id, page_size_bytes)};
-    return CreateCircularBuffer(program, core_range_set, cb_config);
-}
-
-KernelHandle create_kernel(
-    Program& program,
-    const char* kernel_path,
-    const CoreRangeSet& core_range_set,
-    const std::variant<DataMovementConfig, ComputeConfig, EthernetConfig>& config,
-    const std::vector<uint32_t>& runtime_args = {}) {
-    auto kernel_id{CreateKernel(program, kernel_path, core_range_set, config)};
-
-    if (!runtime_args.empty()) {
-        SetRuntimeArgs(program, kernel_id, core_range_set, runtime_args);
-    }
-
-    return kernel_id;
-}
 
 ScatterProgramFactory::cached_program_t ScatterProgramFactory::create(
     const operation_attributes_t& args, const tensor_args_t& tensor_args, tensor_return_value_t& output_tensor) {
