+++ ttnn/cpp/ttnn/operations/data_movement/scatter/device/scatter_program_factory.hpp
@@ -13,13 +13,6 @@
 
 namespace ttnn::operations::data_movement::scatter {
 
-enum class ScatterCB : std::underlying_type_t<tt::CBIndex> {
-    INPUT = CBIndex::c_0,
-    SRC = CBIndex::c_1,
-    INDEX = CBIndex::c_2,
-    DST = CBIndex::c_3,
-};
-
 using namespace tt;
 using namespace tt::tt_metal;
