+++ ttnn/cpp/ttnn/operations/data_movement/scatter/device/scatter_reduce_bfloat16_program_factory.cpp
@@ -4,6 +4,8 @@
 
 #include "scatter_reduce_bfloat16_program_factory.hpp"
 
+#include "scatter_common.hpp"
+
 #include "scatter_device_operation_types.hpp"
 #include "tt-metalium/device.hpp"
 
@@ -12,61 +14,8 @@
 
 namespace ttnn::operations::data_movement::scatter {
 
-namespace {
 using namespace tt;
 using namespace tt::tt_metal;
-constexpr uint32_t BIT_MASK_32 = 32 - 1;
-
-uint64_t ceil32(const uint64_t& number) {
-    return ((number & BIT_MASK_32) == 0) ? number : ((number | BIT_MASK_32) + 1);
-}
-
-}  // namespace
-
-inline uint32_t get_max_l1_space(IDevice* device) {
-    auto lowest_address = device->lowest_occupied_compute_l1_address();
-    uint32_t max_l1_space = lowest_address.has_value() ? lowest_address.value() : device->l1_size_per_core();
-    max_l1_space = max_l1_space - device->allocator()->get_base_allocator_addr(HalMemType::L1);
-    return max_l1_space;
-}
-
-// maximal input/index/source/output chunk size, divisible by 32, calculated as follows:
-// BH available L1 mem size of nearly 1.5 MB...
-// ... divided by 5 to be able to allocate five equally long row chunks (coming from input/index/source/output
-// tensors)
-// ... divided by 4 to account for 4-byte datum sizes of each tensor (fp32, int32)
-// ... minimized by ~10% to account for reserved memory
-inline uint32_t calculate_optimal_chunk_size(IDevice* device) {
-    return ceil32(((((get_max_l1_space(device)) / 5) / 4) * 0.9) - 32);
-}
-
-static CBHandle create_cb(
-    Program& program,
-    const DataType& dtype,
-    const ScatterReduceBfloat16CB& scatter_cb,
-    const CoreRangeSet& core_range_set,
-    const uint32_t& page_size_bytes) {
-    const uint32_t cb_id{static_cast<uint32_t>(scatter_cb)};
-    const auto cb_data_format{datatype_to_dataformat_converter(dtype)};
-    const auto cb_config{
-        CircularBufferConfig{page_size_bytes, {{cb_id, cb_data_format}}}.set_page_size(cb_id, page_size_bytes)};
-    return CreateCircularBuffer(program, core_range_set, cb_config);
-}
-
-static KernelHandle create_kernel(
-    Program& program,
-    const char* kernel_path,
-    const CoreRangeSet& core_range_set,
-    const std::variant<DataMovementConfig, ComputeConfig, EthernetConfig>& config,
-    const std::vector<uint32_t>& runtime_args = {}) {
-    auto kernel_id{CreateKernel(program, kernel_path, core_range_set, config)};
-
-    if (!runtime_args.empty()) {
-        SetRuntimeArgs(program, kernel_id, core_range_set, runtime_args);
-    }
-
-    return kernel_id;
-}
 
 ScatterReduceBfloat16ProgramFactory::cached_program_t ScatterReduceBfloat16ProgramFactory::create(
     const operation_attributes_t& args, const tensor_args_t& tensor_args, tensor_return_value_t& output_tensor) {
@@ -110,7 +59,7 @@ ScatterReduceBfloat16ProgramFactory::cached_program_t ScatterReduceBfloat16Progr
     // ... divided by 5 to be able to allocate five equally long row chunks (coming from input/index/source/output
     // tensors)
     // ... divided by 4 to account for 4-byte datum sizes of each tensor (fp32, int32)
-    // ... minimized by ~20% to account for reserved memory
+    // ... minimized by 120% to account for reserved memory
     const uint32_t input_and_output_max_chunk_size = calculate_optimal_chunk_size(input_tensor.device());
     const uint32_t index_and_source_max_chunk_size = input_and_output_max_chunk_size;
     const uint32_t input_and_output_chunk_size = std::min(input_stick_size, input_and_output_max_chunk_size);
@@ -138,11 +87,11 @@ ScatterReduceBfloat16ProgramFactory::cached_program_t ScatterReduceBfloat16Progr
         index_tensor.buffer()->address(),
         src_tensor.buffer()->address(),
         output_tensor.buffer()->address(),
-        static_cast<uint32_t>(ScatterReduceBfloat16CB::INPUT),
-        static_cast<uint32_t>(ScatterReduceBfloat16CB::INDEX),
-        static_cast<uint32_t>(ScatterReduceBfloat16CB::SRC),
-        static_cast<uint32_t>(ScatterReduceBfloat16CB::DST),
-        static_cast<uint32_t>(ScatterReduceBfloat16CB::FP32_TEMP),
+        static_cast<uint32_t>(ScatterCB::INPUT),
+        static_cast<uint32_t>(ScatterCB::INDEX),
+        static_cast<uint32_t>(ScatterCB::SRC),
+        static_cast<uint32_t>(ScatterCB::DST),
+        static_cast<uint32_t>(ScatterCB::FP32_TEMP),
         input_stick_size,
         index_stick_size,
         source_stick_size,
@@ -170,11 +119,11 @@ ScatterReduceBfloat16ProgramFactory::cached_program_t ScatterReduceBfloat16Progr
     const auto farthest_x_y =
         args.sub_core_grid.has_value() ? args.sub_core_grid->bounding_box().end_coord : compute_with_storage_grid_size;
     const uint32_t all_cores_in_bounding_box = (farthest_x_y.x + 1) * (farthest_x_y.y + 1);
-    create_cb(program, input_tensor.dtype(), ScatterReduceBfloat16CB::INPUT, all_cores, input_page_size_bytes);
-    create_cb(program, index_tensor.dtype(), ScatterReduceBfloat16CB::INDEX, all_cores, index_page_size_bytes);
-    create_cb(program, src_tensor.dtype(), ScatterReduceBfloat16CB::SRC, all_cores, source_page_size_bytes);
-    create_cb(program, output_tensor.dtype(), ScatterReduceBfloat16CB::DST, all_cores, output_page_size_bytes);
-    create_cb(program, fp32_temp_dtype, ScatterReduceBfloat16CB::FP32_TEMP, all_cores, fp32_temp_page_size_bytes);
+    create_cb(program, input_tensor.dtype(), ScatterCB::INPUT, all_cores, input_page_size_bytes);
+    create_cb(program, index_tensor.dtype(), ScatterCB::INDEX, all_cores, index_page_size_bytes);
+    create_cb(program, src_tensor.dtype(), ScatterCB::SRC, all_cores, source_page_size_bytes);
+    create_cb(program, output_tensor.dtype(), ScatterCB::DST, all_cores, output_page_size_bytes);
+    create_cb(program, fp32_temp_dtype, ScatterCB::FP32_TEMP, all_cores, fp32_temp_page_size_bytes);
 
     auto reader_kernel =
         create_kernel(program, reader_kernel_path, all_cores, ReaderDataMovementConfig{compile_time_args});
